# SAMPLE-GO-APP/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "go-duck API",
    "version": "1.0.0",
    "description": "Generated documentation for go-duck microservice"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    }
  ],
  "paths": {
    "/api/cars": {
      "post": {
        "summary": "Create a new Car ",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Car"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          }
        },
        "tags": [
          "Car"
        ]
      },
      "get": {
        "summary": "Get all Cars ",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Zero-based page index"
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Records per page"
          },
          {
            "name": "eager",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "If true, performs SQL Join to fetch relations"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Car"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Car"
        ]
      }
    },
    "/api/cars/{id}": {
      "get": {
        "summary": "Get Car by ID ",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "eager",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          }
        },
        "tags": [
          "Car"
        ]
      },
      "put": {
        "summary": "Update Car ",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          }
        },
        "tags": [
          "Car"
        ]
      },
      "patch": {
        "summary": "Patch Car ",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patched"
          }
        },
        "tags": [
          "Car"
        ]
      },
      "delete": {
        "summary": "Delete Car ",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "Car"
        ]
      }
    },
    "/api/cars/bulk": {
      "post": {
        "summary": "Bulk Create Cars ",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "Car"
        ]
      },
      "put": {
        "summary": "Bulk Update Cars ",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          }
        },
        "tags": [
          "Car"
        ]
      },
      "patch": {
        "summary": "Bulk Patch Cars ",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "changes": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Patched"
          }
        },
        "tags": [
          "Car"
        ]
      }
    },
    "/open/api/cars": {
      "get": {
        "summary": "Get all Cars (Public)",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Zero-based page index"
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Records per page"
          },
          {
            "name": "eager",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "If true, performs SQL Join to fetch relations"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Car"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Car (Public)"
        ],
        "security": [
          {
            "TenantID": []
          }
        ]
      }
    },
    "/open/api/cars/{id}": {
      "get": {
        "summary": "Get Car by ID (Public)",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "eager",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          }
        },
        "tags": [
          "Car (Public)"
        ],
        "security": [
          {
            "TenantID": []
          }
        ]
      },
      "put": {
        "summary": "Update Car (Public)",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          }
        },
        "tags": [
          "Car (Public)"
        ],
        "security": [
          {
            "TenantID": []
          }
        ]
      },
      "patch": {
        "summary": "Patch Car (Public)",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patched"
          }
        },
        "tags": [
          "Car (Public)"
        ],
        "security": [
          {
            "TenantID": []
          }
        ]
      }
    },
    "/open/api/cars/bulk": {
      "put": {
        "summary": "Bulk Update Cars (Public)",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          }
        },
        "tags": [
          "Car (Public)"
        ],
        "security": [
          {
            "TenantID": []
          }
        ]
      },
      "patch": {
        "summary": "Bulk Patch Cars (Public)",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "changes": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Patched"
          }
        },
        "tags": [
          "Car (Public)"
        ],
        "security": [
          {
            "TenantID": []
          }
        ]
      }
    },
    "/api/persons": {
      "post": {
        "summary": "Create a new Person ",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          }
        },
        "tags": [
          "Person"
        ]
      },
      "get": {
        "summary": "Get all Persons ",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Zero-based page index"
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Records per page"
          },
          {
            "name": "eager",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "If true, performs SQL Join to fetch relations"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Person"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Person"
        ]
      }
    },
    "/api/persons/{id}": {
      "get": {
        "summary": "Get Person by ID ",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "eager",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          }
        },
        "tags": [
          "Person"
        ]
      },
      "put": {
        "summary": "Update Person ",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          }
        },
        "tags": [
          "Person"
        ]
      },
      "patch": {
        "summary": "Patch Person ",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patched"
          }
        },
        "tags": [
          "Person"
        ]
      },
      "delete": {
        "summary": "Delete Person ",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "Person"
        ]
      }
    },
    "/api/persons/bulk": {
      "post": {
        "summary": "Bulk Create Persons ",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "Person"
        ]
      },
      "put": {
        "summary": "Bulk Update Persons ",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          }
        },
        "tags": [
          "Person"
        ]
      },
      "patch": {
        "summary": "Bulk Patch Persons ",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "changes": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Patched"
          }
        },
        "tags": [
          "Person"
        ]
      }
    },
    "/open/api/persons": {
      "get": {
        "summary": "Get all Persons (Public)",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Zero-based page index"
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Records per page"
          },
          {
            "name": "eager",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "If true, performs SQL Join to fetch relations"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Person"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Person (Public)"
        ],
        "security": [
          {
            "TenantID": []
          }
        ]
      }
    },
...
    "/audit": {
      "get": {
        "tags": [
          "Observability"
        ],
        "summary": "Fetch Audit Trail",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "default"
            },
            "description": "Multi-tenancy context identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "TenantID": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tenant-ID",
        "description": "The unique identifier for the tenant context"
      }
    },
    "schemas": {
...
      "Car": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "year": {
            "type": "integer"
          },
          "price": {
            "type": "number"
          },
          "color": {
            "type": "string"
          },
          "features": {
            "type": "object"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": [],
      "TenantID": []
    }
  ]
}